+++ src/main/java/gtexpert/integration/gendustry/metatileentities/MetaTileEntityIndustrialApiary.java
@@ -17,7 +17,6 @@ import gregtech.api.capability.impl.NotifiableItemStackHandler;
 import gregtech.api.capability.impl.RecipeLogicEnergy;
 import gregtech.api.metatileentity.MetaTileEntity;
 import gregtech.api.metatileentity.interfaces.IGregTechTileEntity;
-import gregtech.api.recipes.RecipeMap;
 import gregtech.client.renderer.ICubeRenderer;
 
 import gtexpert.common.metatileentities.GTESimpleMachineMetaTileEntity;