+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# v2.8.9
+- Fix Implosion Compressor recipes.
+
+* * *
+
+# v2.2.8
+- Electric Implosion Compressor is converted to no explosive items. [#284](https://github.com/GTModpackTeam/GTExpert-Core/pull/284)
+    - Thanks to @MrKono and @sysnote8main for his support!! 
+
+* * *
+
 # v2.2.7
 - Fix Ender Pearl and Ender Eye recipes.
 - Fix LCU(Large Cracking Unit).